+++ resources-ktx/CHANGELOG.md
@@ -2,3 +2,4 @@
 ### Added
 - **View**, **Fragment**: new extensions `getDimensionPixelSize` and `getDimensionPixelOffset`
 - **Context**: new extensions `resolveAttributeOrThrow` and `resolveAttribute`
+- **Context**: new extensions `resolveColor`